refactor: reduce template event action surface
